requirements/local.txt
-r base.txt

watchdog==2.0.2
Werkzeug==1.0.1
django-extensions==3.1.3

factory_boy==3.2.0
faker==8.1.0
parameterized==0.8.1
django-debug-toolbar==3.2.1

django-rosetta==0.9.6

Sphinx==3.5.4
sphinx-press-theme==0.5.1
